Artwork:
Provide two icons: a generic club-mate icon (which is the same as the icon we 
used so far) and an 'obsolete' icon which is the mate-logo with a 
attention-sign. The machMatekate.rb script was updated to use these icons. I 
addition, the icons are now referenced with a relative path (instead of an 
absolute one).


git-svn-id: https://svn.cccmz.de/matekate@19 223dcf81-e940-4002-9f84-9aebb59a28bd